Refactor AudienceTilesWidget to be a JS module.
